refactoring and amending
