+++ spartacus/__init__.py
@@ -1,19 +1,5 @@
-from .enums import CartesianAxis, EulerSequence, JointType, DatasetCSV, BiomechDirection, BiomechOrigin, Segment
-from .utils import (
-    biomech_direction_string_to_enum,
-    biomech_origin_string_to_enum,
-    biomech_direction_sign,
-    BiomechCoordinateSystem,
-    Joint,
-    joint_string_to_enum,
-    euler_sequence_to_enum,
-    segment_str_to_enum,
-    get_segment_columns,
-    get_correction_column,
-    get_conversion_from_not_isb_to_isb_oriented,
-    convert_rotation_matrix_from_one_coordinate_system_to_another,
-)
-from .checks import (
+from src.enums import CartesianAxis, EulerSequence, JointType, DatasetCSV, BiomechDirection, BiomechOrigin, Segment
+from src.checks import (
     check_parent_child_joint,
     check_segment_filled_with_nan,
     check_is_isb_segment,
@@ -22,5 +8,5 @@ from .checks import (
     check_same_orientation,
 )
 
-from .row_data import RowData
-from .load import load
+from .src.row_data import RowData
+from .src.load import load
