+++ spartacus/src/enums.py
