+++ spartacus/src/utils.py
